refactor: 非同期処理を呼び出す関数をapiFunc.tsにて宣言
